Minor adjustment and unit test

+++ ugs-core/src/com/willwinder/universalgcodesender/firmware/g2core/G2CoreOverrideManager.java
@@ -52,7 +52,7 @@ public class G2CoreOverrideManager extends AbstractOverrideManager implements IO
     }
 
     @Override
-    protected float getSpeedMajorStep(OverrideType overrideType) {
+    protected int getSpeedMajorStep(OverrideType overrideType) {
         return MAJOR_STEP;
     }
